requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv